Remove last update timestamp from completion file
The last updated timestamp is not so useful as it is hard to maintain
the same timestamp on both the completion script and the git commit.
Also git makes a note of the commit timestamp while commiting the
changes, so to know when the completion script was last updated we can
get to know from `git blame`.